requirements.txt
fastapi[standard]
pyjwt 
pymongo 